Enhance database schema and user authentication: add ON DELETE behavior for foreign keys, implement password verification in user operations, and introduce triggers for user deletion handling.

+++ server/AccountHandler.java
@@ -2,6 +2,7 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.sql.ResultSet;
 
 public class AccountHandler {
     private Connection conn;
@@ -20,12 +21,16 @@ public class AccountHandler {
     }
 
     public User authenticateUser(String username, String password) throws SQLException {
+        if(!password.equals(getPassword(username))) {
+            throw new SQLException("Password incorrect!");
+        }
+
         String sql = "SELECT name, email FROM Users WHERE name = ? AND password = ?";
         PreparedStatement ps = conn.prepareStatement(sql);
         ps.setString(1, username);
         ps.setString(2, password);
         ResultSet rs = ps.executeQuery();
-        if(rs.next()) {
+        if (rs.next()) {
             return new User(username, rs.getString("email"), password);
         }
         else {
@@ -34,9 +39,27 @@ public class AccountHandler {
     }
 
     public void deleteUser(String username, String password) throws SQLException {
+        if(!password.equals(getPassword(username))) {
+            throw new SQLException("Password incorrect!");
+        }
+
         String sql = "DELETE FROM Users WHERE name = ? AND password = ?";
         PreparedStatement ps = conn.prepareStatement(sql);
         ps.setString(1, username);
         ps.setString(2, password);
+        ps.executeUpdate();
+    }
+
+    private String getPassword(String username) throws SQLException {
+        String sql = "SELECT password FROM Users WHERE name = ?";
+        PreparedStatement ps = conn.prepareStatement(sql);
+        ps.setString(1, username);
+        ResultSet rs = ps.executeQuery();
+        if(rs.next()) {
+            return rs.getString("password");
+        }
+        else  {
+            throw new SQLException(String.format("The user %d does not exist!", username));
+        }
     }
 }
